[issue-137] Moved Serializers to correct app location
