+++ cocoon/houseDatabase/serializers.py
@@ -1,0 +1,39 @@
+# Import DRF modules
+from rest_framework import serializers
+
+# Import App modules
+from .models import RentDatabaseModel, HomeTypeModel, HousePhotos
+
+
+class HomeImageSerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = HousePhotos
+        fields = ('id', 'image')
+
+
+class HomeTypeSerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = HomeTypeModel
+        fields = ('home_type',)
+
+
+class FavoritesSerializer(serializers.HyperlinkedModelSerializer):
+    home_type = HomeTypeSerializer(read_only=True)
+    images = HomeImageSerializer(read_only=True, many=True)
+
+    class Meta:
+        model = RentDatabaseModel
+        fields = ('id', 'price', 'home_type', 'images')
+
+
+class VisitListSerializer(serializers.HyperlinkedModelSerializer):
+    home_type = HomeTypeSerializer(read_only=True)
+    images = HomeImageSerializer(read_only=True, many=True)
+
+    class Meta:
+        model = RentDatabaseModel
+        fields = ('id', 'price', 'home_type', 'images')
+
+
