+++ cocoon/survey/serializers.py
@@ -1,39 +1,11 @@
+# Import DRF modules
 from rest_framework import serializers
 
-from cocoon.survey.models import RentingSurveyModel
-from cocoon.houseDatabase.models import RentDatabaseModel, HomeTypeModel, HousePhotos
+# Import App modules
+from .models import RentingSurveyModel
 
-
-class HomeImageSerializer(serializers.HyperlinkedModelSerializer):
-
-    class Meta:
-        model = HousePhotos
-        fields = ('id', 'image')
-
-
-class HomeTypeSerializer(serializers.HyperlinkedModelSerializer):
-
-    class Meta:
-        model = HomeTypeModel
-        fields = ('home_type',)
-
-
-class FavoritesSerializer(serializers.HyperlinkedModelSerializer):
-    home_type = HomeTypeSerializer(read_only=True)
-    images = HomeImageSerializer(read_only=True, many=True)
-
-    class Meta:
-        model = RentDatabaseModel
-        fields = ('id', 'price', 'home_type', 'images')
-
-
-class VisitListSerializer(serializers.HyperlinkedModelSerializer):
-    home_type = HomeTypeSerializer(read_only=True)
-    images = HomeImageSerializer(read_only=True, many=True)
-
-    class Meta:
-        model = RentDatabaseModel
-        fields = ('id', 'price', 'home_type', 'images')
+# Import Third party modules
+from cocoon.houseDatabase.serializers import FavoritesSerializer, VisitListSerializer
 
 
 class RentSurveySerializer(serializers.HyperlinkedModelSerializer):
